Move RtpCodecCapabiltiy into separate module
