+++ crates/native/src/api/mod.rs
@@ -4,6 +4,7 @@ pub mod media_device_info;
 pub mod media_display_info;
 pub mod rtc_ice_candidate_stats;
 pub mod rtp_capabilities;
+pub mod rtp_codec_capability;
 
 use std::{
     sync::{
@@ -27,6 +28,7 @@ pub use self::{
         RtpCapabilities, get_rtp_receiver_capabilities,
         get_rtp_sender_capabilities,
     },
+    rtp_codec_capability::{RtpCodecCapability, set_codec_preferences},
 };
 // Re-exporting since it is used in the generated code.
 pub use crate::{
@@ -770,82 +772,6 @@ impl From<sys::RtpHeaderExtensionCapability> for RtpHeaderExtensionCapability {
             preferred_id: value.preferred_id(),
             preferred_encrypted: value.preferred_encrypted(),
             direction: value.direction().into(),
-        }
-    }
-}
-
-/// Representation of static capabilities of an endpoint's implementation of a
-/// codec.
-#[derive(Debug)]
-pub struct RtpCodecCapability {
-    /// Default payload type for the codec.
-    ///
-    /// Mainly needed for codecs that have statically assigned payload types.
-    pub preferred_payload_type: Option<i32>,
-
-    /// List of [`ScalabilityMode`]s supported by the video codec.
-    pub scalability_modes: Vec<ScalabilityMode>,
-
-    /// Built [MIME "type/subtype"][0] string from `name` and `kind`.
-    ///
-    /// [0]: https://en.wikipedia.org/wiki/Media_type
-    pub mime_type: String,
-
-    /// Used to identify the codec. Equivalent to [MIME subtype][0].
-    ///
-    /// [0]: https://en.wikipedia.org/wiki/Media_type#Subtypes
-    pub name: String,
-
-    /// [`MediaType`] of this codec. Equivalent to [MIME] top-level type.
-    ///
-    /// [MIME]: https://en.wikipedia.org/wiki/Media_type
-    pub kind: MediaType,
-
-    /// If [`None`], the implementation default is used.
-    pub clock_rate: Option<i32>,
-
-    /// Number of audio channels used.
-    ///
-    /// [`None`] for video codecs.
-    ///
-    /// If [`None`] for audio, the implementation default is used.
-    pub num_channels: Option<i32>,
-
-    /// Codec-specific parameters that must be signaled to the remote party.
-    ///
-    /// Corresponds to `a=fmtp` parameters in [SDP].
-    ///
-    /// Contrary to ORTC, these parameters are named using all lowercase
-    /// strings. This helps make the mapping to [SDP] simpler, if an application
-    /// is using [SDP]. Boolean values are represented by the string "1".
-    ///
-    /// [SDP]: https://en.wikipedia.org/wiki/Session_Description_Protocol
-    pub parameters: Vec<(String, String)>,
-
-    /// Feedback mechanisms to be used for this codec.
-    pub feedback: Vec<RtcpFeedback>,
-}
-
-impl From<sys::RtpCodecCapability> for RtpCodecCapability {
-    fn from(value: sys::RtpCodecCapability) -> Self {
-        Self {
-            preferred_payload_type: value.preferred_payload_type(),
-            scalability_modes: value
-                .scalability_modes()
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-            mime_type: value.mime_type(),
-            name: value.name(),
-            kind: value.kind().into(),
-            clock_rate: value.clock_rate(),
-            num_channels: value.num_channels(),
-            parameters: value.parameters().into_iter().collect(),
-            feedback: value
-                .rtcp_feedback()
-                .into_iter()
-                .map(Into::into)
-                .collect(),
         }
     }
 }
@@ -2580,16 +2506,6 @@ pub fn stop_transceiver(
     transceiver: RustOpaque<Arc<RtpTransceiver>>,
 ) -> anyhow::Result<()> {
     transceiver.stop()
-}
-
-/// Changes the preferred [`RtpTransceiver`] codecs to the provided
-/// [`Vec`]`<`[`RtpCodecCapability`]`>`.
-#[expect(clippy::needless_pass_by_value, reason = "FFI")]
-pub fn set_codec_preferences(
-    transceiver: RustOpaque<Arc<RtpTransceiver>>,
-    codecs: Vec<RtpCodecCapability>,
-) {
-    transceiver.set_codec_preferences(codecs);
 }
 
 /// Replaces the specified [`AudioTrack`] (or [`VideoTrack`]) on the
