+++ lib/src/api/bridge/frb.dart
@@ -3,3 +3,4 @@ export 'api/media_device_info.dart';
 export 'api/media_display_info.dart';
 export 'api/rtc_ice_candidate_stats.dart';
 export 'api/rtp_capabilities.dart';
+export 'api/rtp_codec_capability.dart';
